@@ -10,6 +10,7 @@ import * as newsStore from '../../../news/store';
 import * as fromRoot from '../../../store';
 import {Topic} from '../../../topics/types/topic';
 import {Logout} from '../../../store';
+import {NewsFeedbackComponent} from '../news-feedback/news-feedback.component';
 
 @Component({
   selector: 'top-bar',
@@ -26,9 +27,13 @@ import {Logout} from '../../../store';
                     (view)="onView($event)"></audio-player>
       <div class="user-menu">
         <button mat-icon-button class="main-button" [matMenuTriggerFor]="userMenu">
-          <mat-icon class="big-icon">account_circle</mat-icon>
+          <mat-icon class="big-icon">dehaze</mat-icon>
         </button>
         <mat-menu #userMenu="matMenu">
+          <button mat-menu-item (click)="onFeedback()">
+            <mat-icon>feedback</mat-icon>
+            <span>Feedback</span>
+          </button>
           <button mat-menu-item (click)="onTopics()">
             <mat-icon>list</mat-icon>
             <span>Topics</span>
@@ -49,7 +54,7 @@ export class TopBarComponent implements OnInit {
   currentNews$: Observable<NewsEntry>;
   playing$: Observable<boolean>;
   topics$: Observable<Topic[]>;
-  audio$: Observable<HTMLAudioElement>
+  audio$: Observable<HTMLAudioElement>;
 
   constructor(
     private store: Store<newsStore.NewsState>,
@@ -94,9 +99,18 @@ export class TopBarComponent implements OnInit {
       panelClass: 'full-screen-modal',
     };
     const dialogRef = this.dialog.open(TopicsPageComponent, {
-        panelClass: 'my-test-class'
+        panelClass: 'full-popup'
       }
     );
+
+    dialogRef.afterClosed().subscribe(result => {
+    });
+  }
+
+  onFeedback() {
+    const dialogRef = this.dialog.open(NewsFeedbackComponent, {
+      panelClass: 'full-popup'
+    });
 
     dialogRef.afterClosed().subscribe(result => {
     });
